+++ src/Domain/Repository/TypeRepositoryInterface.php
@@ -186,9 +186,9 @@ interface TypeRepositoryInterface
      * @param Event  $event
      * @param string $locale
      *
-     * @return Type[]
+     * @return TypeView[]
      */
-    public function getVisibleTypesViewsByEvent(Event $event, $locale);
+    public function getVisibleTypesViewsByEvent(Event $event, $locale): array;
 
     public function hasVisibleTypeByEvent(Event $event): bool;